feat: refactor to use common savesecretkey hook
